move performance tests to their own class
